% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_fraction.R
\name{insights_fraction}
\alias{insights_fraction}
\title{Apply InSiGHTS with fractional land use data}
\usage{
insights_fraction(range, lu, other, outfile = NULL, clamp = FALSE)

\S4method{insights_fraction}{SpatRaster,SpatRaster,SpatRaster,character,logical}(range,lu,other,outfile,clamp)

\S4method{insights_fraction}{SpatRaster,stars,ANY,character,logical}(range,lu,other,outfile,clamp)

\S4method{insights_fraction}{stars,stars,ANY,character,logical}(range,lu,other,outfile,clamp)

\S4method{insights_fraction}{stars,SpatRaster,ANY,character,logical}(range,lu,other,outfile,clamp)

\S4method{insights_fraction}{ANY,ANY,SpatRaster,character,logical}(range,lu,other,outfile,clamp)
}
\arguments{
\item{range}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} or temporal \code{\link{stars}} object describing the estimated distribution of a
biodiversity feature (e.g. species). \strong{Has to be in binary format!}
Alternatively a \code{DistributionModel} fitted with \code{ibis.iSDM} package can be supplied.}

\item{lu}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} or temporal \code{\link{stars}} object of the future land-use fractions to be applied to the range.
\strong{Each layer has to be in units of fractions, e.g. between 0 and 1.}}

\item{other}{Any other \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} or temporal \code{\link{stars}} objects that describe suitable conditions for the species.}

\item{outfile}{A writeable \code{\link{character}} of where the output should be written to. If missing, the function will return
a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} or \code{\link{stars}} object respectively.}

\item{clamp}{A \code{\link{logical}} on whether lu should be clamped to 0 and 1 beforehand (Default: \code{FALSE}).}
}
\value{
Either a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} or temporal \code{\link{stars}} object or nothing if outputs are written directly to drive.
}
\description{
This function applies an area-of-habitat (AOH) correction to a provided single time step or multiple step range estimate
(binary format). This thus assumes that species-habitat relations remain stable also in future conditions within a
provided climatic niche.

It is assumed that the land-use layers come in fractional units, so are ranging from \code{0} to \code{1}.
Optionally also a elevation (\code{elev}) layer and habitat condition (\code{condition}) can be provided to support refinements
by elevational range or habitat condition.
}
\note{
This function does not do the refinement of land-use fraction to relevant habitats.
This needs to be done by the analyst a-priori.
}
\examples{
\dontrun{
 out <- insights_fraction(range, landuse)
}

}
\references{
\itemize{
\item Rondinini, Carlo, and Piero Visconti. "Scenarios of large mammal loss in Europe for the 21st century." Conservation Biology 29, no. 4 (2015): 1028-1036.
\item Visconti, Piero, Michel Bakkenes, Daniele Baisero, Thomas Brooks, Stuart HM Butchart, Lucas Joppa, Rob Alkemade et al. "Projecting global biodiversity indicators under future development scenarios." Conservation Letters 9, no. 1 (2016): 5-13.
}
}
\author{
Martin Jung
}
