% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_area.R
\name{insights_area}
\alias{insights_area}
\title{Apply InSiGHTS with continuous areal land use data}
\usage{
insights_area(range, lu, other, outfile = NULL)

\S4method{insights_area}{SpatRaster,SpatRaster,SpatRaster,character}(range,lu,other,outfile)

\S4method{insights_area}{SpatRaster,stars,ANY,character}(range,lu,other,outfile)

\S4method{insights_area}{stars,stars,ANY,character}(range,lu,other,outfile)

\S4method{insights_area}{stars,SpatRaster,ANY,character}(range,lu,other,outfile)

\S4method{insights_area}{ANY,ANY,SpatRaster,character}(range,lu,other,outfile)
}
\arguments{
\item{range}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} or temporal \code{\link{stars}} object describing the estimated distribution of a
biodiversity feature (e.g. species). Alternatively a \code{DistributionModel} fitted with \code{ibis.iSDM} package can be supplied.}

\item{lu}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} or temporal \code{\link{stars}} object of the future land-use areas to be applied to the range.
\strong{Each layer has to be in units of area, so greater 0}}

\item{other}{Any other \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} or temporal \code{\link{stars}} objects that describe suitable conditions for the species.}

\item{outfile}{A writeable \code{\link{character}} of where the output should be written to. If missing, the function will return
a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} or \code{\link{stars}} object respectively.}
}
\value{
Either a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} or temporal \code{\link{stars}} object or nothing if outputs are written directly to drive.
}
\description{
This function applies an area-of-habitat (AOH) correction to a provided single time step or multiple step range estimate
(binary or other format). This thus assumes that species-habitat relations remain stable also in future conditions within a
provided climatic niche.

Opposed to \code{\link{insights_fraction}} it is assumed that the land-use layers come in continuous areal units,
such as \code{'km2'}.
Optionally also a elevation (\code{elev}) layer and habitat condition (\code{condition}) can be provided to support refinements
by elevational range or habitat condition.
}
\note{
This function does not do the refinement of land-use fraction to relevant habitats.
This needs to be done by the analyst a-priori.
\emph{No checks are conducted on whether area shares match or exceed certain amounts!}
}
\examples{
\dontrun{
 out <- insights_area(range, landuse_km2)
}

}
\references{
\itemize{
\item Rondinini, Carlo, and Piero Visconti. "Scenarios of large mammal loss in Europe for the 21st century." Conservation Biology 29, no. 4 (2015): 1028-1036.
\item Visconti, Piero, Michel Bakkenes, Daniele Baisero, Thomas Brooks, Stuart HM Butchart, Lucas Joppa, Rob Alkemade et al. "Projecting global biodiversity indicators under future development scenarios." Conservation Letters 9, no. 1 (2016): 5-13.
}
}
\author{
Martin Jung
}
